Move vocabs from asr to common (#3084)
* move vocabs to common


* fix style


* move vocab import to the function scope
